+++ de.bund.bfr.knime.fsklab.nodes/META-INF/MANIFEST.MF
@@ -76,5 +76,6 @@ Export-Package: de.bund.bfr.knime.fsklab;
    javax.swing,
    de.bund.bfr.knime.fsklab.r.client",
  de.bund.bfr.knime.fsklab.nodes.environment,
- de.bund.bfr.knime.fsklab.nodes.plot;uses:="org.knime.python2.kernel,de.bund.bfr.knime.fsklab.r.client"
+ de.bund.bfr.knime.fsklab.nodes.plot;uses:="org.knime.python2.kernel,de.bund.bfr.knime.fsklab.r.client",
+ de.bund.bfr.knime.fsklab.v1_9.runner
 Import-Package: org.w3c.dom.events;resolution:=optional
