+++ de.bund.bfr.knime.fsklab.nodes/src/de/bund/bfr/knime/fsklab/nodes/PythonScriptHandler.java
@@ -11,6 +11,8 @@ import org.knime.python2.kernel.PythonKernelOptions;
 import de.bund.bfr.knime.fsklab.FskPortObject;
 import de.bund.bfr.knime.fsklab.FskSimulation;
 import de.bund.bfr.knime.fsklab.nodes.plot.PythonPlotter;
+import de.bund.bfr.knime.fsklab.v1_9.runner.RunnerNodeInternalSettings;
+import de.bund.bfr.knime.fsklab.v1_9.runner.RunnerNodeSettings;
 
 public class PythonScriptHandler extends ScriptHandler {
   String std_out = "";
